Split utils.py into sampling, extraction, clustering modules
- Replace utils.py with three focused modules:
  - sampling.py: trace sampling, session grouping, scorer verification
  - extraction.py: execution path, span errors, failing traces, failure labels
  - clustering.py: LLM clustering, summarization, build_summary, log_artifacts
- Inline cluster_analyses() into pipeline.py (trivial wrapper, single call site)
- Split test_utils.py into matching test files
- Simplify constants (consolidate max_tokens, add SAMPLE_POOL_MULTIPLIER)
- Add section comments to extract_span_errors for readability
