+++ mlflow/genai/discovery/constants.py
@@ -4,15 +4,10 @@ from typing import Literal
 
 DEFAULT_TRIAGE_SAMPLE_SIZE = 100
 MAX_EXAMPLE_TRACE_IDS = 10
-MIN_EXAMPLES = 1
+SAMPLE_POOL_MULTIPLIER = 5
 SAMPLE_RANDOM_SEED = 42
 NUM_RETRIES = 5
-
-# LLM inference parameter limits
-LABEL_MAX_TOKENS = 1000
-CLUSTER_MAX_TOKENS = 8000
-SUMMARY_MAX_TOKENS = 4096
-ANNOTATION_MAX_TOKENS = 4096
+LLM_MAX_TOKENS = 8192
 
 # Text truncation limits
 SURFACE_TRUNCATION_LIMIT = 800
