+++ mlflow/genai/discovery/pipeline.py
@@ -13,16 +13,21 @@ from mlflow.entities.assessment import Feedback
 from mlflow.entities.assessment_source import AssessmentSource, AssessmentSourceType
 from mlflow.entities.trace import Trace
 from mlflow.environment_variables import MLFLOW_GENAI_EVAL_MAX_WORKERS
+from mlflow.genai.discovery.clustering import (
+    build_summary,
+    cluster_by_llm,
+    log_discovery_artifacts,
+    summarize_cluster,
+)
 from mlflow.genai.discovery.constants import (
-    ANNOTATION_MAX_TOKENS,
     CONFIDENCE_ORDER,
     DEFAULT_ANALYSIS_MODEL,
     DEFAULT_JUDGE_MODEL,
     DEFAULT_SCORER_NAME,
     DEFAULT_TRIAGE_SAMPLE_SIZE,
+    LLM_MAX_TOKENS,
     MAX_EXAMPLE_TRACE_IDS,
     MIN_CONFIDENCE,
-    MIN_EXAMPLES,
     NO_ISSUE_KEYWORD,
     NUM_RETRIES,
     SURFACE_TRUNCATION_LIMIT,
@@ -36,20 +41,17 @@ from mlflow.genai.discovery.entities import (
     _ConversationAnalysis,
     _IdentifiedIssue,
 )
-from mlflow.genai.discovery.utils import (
-    build_summary,
-    cluster_analyses,
-    cluster_by_llm,
+from mlflow.genai.discovery.extraction import (
     extract_execution_path,
     extract_execution_paths_for_session,
     extract_failing_traces,
     extract_failure_labels,
     extract_span_errors,
+)
+from mlflow.genai.discovery.sampling import (
     get_session_id,
     group_traces_by_session,
-    log_discovery_artifacts,
     sample_traces,
-    summarize_cluster,
     verify_scorer,
 )
 from mlflow.genai.judges.adapters.litellm_adapter import _invoke_litellm
@@ -165,7 +167,8 @@ def _recluster_singletons(
     max_issues: int,
     token_counter: _TokenCounter | None = None,
 ) -> list[_IdentifiedIssue]:
-    """Re-cluster singleton issues via a second LLM pass to find better groupings.
+    """
+    Re-cluster singleton issues via a second LLM pass to find better groupings.
 
     Args:
         singletons: Single-analysis issues to attempt merging.
@@ -231,7 +234,8 @@ def _annotate_issue_traces(
     session_first_trace: dict[str, str] | None = None,
     token_counter: _TokenCounter | None = None,
 ) -> None:
-    """Log a Feedback assessment for each issue on affected traces.
+    """
+    Log a Feedback assessment for each issue on affected traces.
 
     When session information is provided, logs one annotation per (issue,
     session) on the session's first trace. Otherwise annotates each trace
@@ -303,7 +307,7 @@ def _annotate_issue_traces(
                 num_retries=NUM_RETRIES,
                 response_format=None,
                 include_response_format=False,
-                inference_params={"max_tokens": ANNOTATION_MAX_TOKENS, "temperature": 0},
+                inference_params={"max_tokens": LLM_MAX_TOKENS, "temperature": 0},
             )
             if token_counter is not None:
                 token_counter.track(response)
@@ -356,7 +360,8 @@ def discover_issues(
     max_issues: int = 20,
     filter_string: str | None = None,
 ) -> DiscoverIssuesResult:
-    """Discover quality and operational issues in traces.
+    """
+    Discover quality and operational issues in traces.
 
     Runs a multi-phase pipeline:
     1. **Triage**: Scores traces using the provided scorers (or a default
@@ -567,13 +572,12 @@ def discover_issues(
 
     _logger.info("Phase 3: Clustering analyses into issues...")
     phase_start = time.time()
-    cluster_groups = cluster_analyses(
-        analyses,
-        max_issues,
-        labels=labels,
-        label_model=judge_model,
-        token_counter=token_counter,
-    )
+    if len(analyses) == 1:
+        cluster_groups = [[0]]
+    else:
+        cluster_groups = cluster_by_llm(
+            labels, max_issues, judge_model, token_counter=token_counter
+        )
     cluster_end = time.time()
     _logger.info(
         "Phase 3: Clustering took %.1fs, produced %d clusters",
@@ -644,9 +648,7 @@ def discover_issues(
     identified: list[_IdentifiedIssue] = [
         issue
         for issue in summaries
-        if confidence_gte(issue.confidence, MIN_CONFIDENCE)
-        and len(issue.example_indices) >= MIN_EXAMPLES
-        and not _is_non_issue(issue)
+        if confidence_gte(issue.confidence, MIN_CONFIDENCE) and not _is_non_issue(issue)
     ]
 
     # Merge issues with identical names (case-insensitive), combining their
