+++ tests/genai/discovery/test_extraction.py
@@ -1,0 +1,123 @@
+import pandas as pd
+
+from mlflow.genai.discovery.extraction import (
+    extract_failing_traces,
+    extract_span_errors,
+)
+from mlflow.genai.evaluation.entities import EvaluationResult
+
+# ---- extract_span_errors ----
+
+
+def test_extract_span_errors_with_error_span(make_trace):
+    trace = make_trace(error_span=True)
+    result = extract_span_errors(trace)
+    assert result
+    assert "Connection failed" in result
+
+
+def test_extract_span_errors_no_errors(make_trace):
+    trace = make_trace()
+    result = extract_span_errors(trace)
+    assert result == ""
+
+
+def test_extract_span_errors_truncation(make_trace):
+    trace = make_trace(error_span=True)
+    result = extract_span_errors(trace, max_length=10)
+    assert len(result) <= 10
+
+
+# ---- extract_failing_traces ----
+
+
+def test_extract_failing_traces(make_trace):
+    traces = [make_trace() for _ in range(3)]
+    df = pd.DataFrame(
+        {
+            "satisfaction/value": [True, False, False],
+            "satisfaction/rationale": ["good", "bad response", "incomplete"],
+            "trace": traces,
+        }
+    )
+    eval_result = EvaluationResult(run_id="run-1", metrics={}, result_df=df)
+
+    failing, rationales = extract_failing_traces(eval_result, "satisfaction")
+
+    assert len(failing) == 2
+    assert failing[0].info.trace_id == traces[1].info.trace_id
+    assert failing[1].info.trace_id == traces[2].info.trace_id
+    assert rationales[traces[1].info.trace_id] == "bad response"
+    assert rationales[traces[2].info.trace_id] == "incomplete"
+
+
+def test_extract_failing_traces_with_list_of_scorer_names(make_trace):
+    traces = [make_trace() for _ in range(3)]
+    df = pd.DataFrame(
+        {
+            "satisfaction/value": [True, False, True],
+            "satisfaction/rationale": ["good", "bad response", "good"],
+            "quality/value": [True, True, False],
+            "quality/rationale": ["ok", "ok", "poor quality"],
+            "trace": traces,
+        }
+    )
+    eval_result = EvaluationResult(run_id="run-1", metrics={}, result_df=df)
+
+    failing, rationales = extract_failing_traces(eval_result, ["satisfaction", "quality"])
+
+    assert len(failing) == 2
+    assert failing[0].info.trace_id == traces[1].info.trace_id
+    assert failing[1].info.trace_id == traces[2].info.trace_id
+    assert rationales[traces[1].info.trace_id] == "bad response"
+    assert rationales[traces[2].info.trace_id] == "poor quality"
+
+
+def test_extract_failing_traces_multiple_scorers_fail_same_row(make_trace):
+    traces = [make_trace() for _ in range(2)]
+    df = pd.DataFrame(
+        {
+            "scorer_a/value": [False, True],
+            "scorer_a/rationale": ["reason a", "ok"],
+            "scorer_b/value": [False, True],
+            "scorer_b/rationale": ["reason b", "ok"],
+            "trace": traces,
+        }
+    )
+    eval_result = EvaluationResult(run_id="run-1", metrics={}, result_df=df)
+
+    failing, rationales = extract_failing_traces(eval_result, ["scorer_a", "scorer_b"])
+
+    assert len(failing) == 1
+    assert failing[0].info.trace_id == traces[0].info.trace_id
+    assert "reason a" in rationales[traces[0].info.trace_id]
+    assert "reason b" in rationales[traces[0].info.trace_id]
+
+
+def test_extract_failing_traces_none_result_df():
+    eval_result = EvaluationResult(run_id="run-1", metrics={}, result_df=None)
+    failing, rationales = extract_failing_traces(eval_result, "satisfaction")
+    assert failing == []
+    assert rationales == {}
+
+
+def test_extract_failing_traces_missing_column():
+    df = pd.DataFrame({"other/value": [True]})
+    eval_result = EvaluationResult(run_id="run-1", metrics={}, result_df=df)
+    failing, rationales = extract_failing_traces(eval_result, "satisfaction")
+    assert failing == []
+
+
+def test_extract_failing_traces_no_failures(make_trace):
+    traces = [make_trace()]
+    df = pd.DataFrame(
+        {
+            "satisfaction/value": [True],
+            "satisfaction/rationale": ["good"],
+            "trace": traces,
+        }
+    )
+    eval_result = EvaluationResult(run_id="run-1", metrics={}, result_df=df)
+    failing, rationales = extract_failing_traces(eval_result, "satisfaction")
+    assert failing == []
+    assert rationales == {}
